+++ pkg/vault/reconcile.go
@@ -2,6 +2,7 @@ package vault
 
 import (
 	"fmt"
+	log "github.com/sirupsen/logrus"
 	"strings"
 	"time"
 )
@@ -111,4 +112,27 @@ func ParseDuration(duration string) (time.Duration, error) {
 	}
 
 	return time.ParseDuration(duration)
+}
+
+// DataInSecret compare given data with data stored in the vault secret
+func DataInSecret(data map[string]interface{}, path string) bool {
+	// read desired secret
+	secret := ReadSecret(path)
+	if secret == nil {
+		return false
+	}
+	for k, v := range data {
+		if strings.HasSuffix(k, "ttl") || strings.HasSuffix(k, "period") {
+			dur, err := time.ParseDuration(v.(string))
+			if err != nil {
+				log.WithError(err).WithField("option", k).Fatal("failed to parse duration from data")
+			}
+			v = int64(dur.Seconds())
+		}
+		if fmt.Sprintf("%v", secret.Data[k]) == fmt.Sprintf("%v", v) {
+			continue
+		}
+		return false
+	}
+	return true
 }
